requirements.txt
google-genai
requests
python-frontmatter
